backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
openai==1.12.0
google-generativeai==0.5.0
pydantic[email]==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
aiohttp==3.9.3
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
tenacity==8.2.3
structlog==24.1.0
pytest==7.4.4
pytest-asyncio==0.23.4

